refactor(cli): extract status bar display model
